refactor: extract scenes
